pyproject.toml
[tool.poetry]
name = "icpi"
version = "0.1.0"
description = ""
authors = ["Ethan Brooks <ethanabrooks@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.16.0"
gym = "^0.21.0"
numpy = "^1.22.3"
pandas = "^1.4.3"
altair = "^4.2.0"
GitPython = "^3.1.27"
sweep-logger = {git = "https://github.com/run-tracker/sweep-logger.git", rev = "main"}
vega-charts = {git = "https://github.com/run-tracker/vega-charts.git", rev = "main"}
bsuite = "^0.3.5"
ray = {version = "^1.13.0", extras = ["tune"]}
dollar-lambda = {git = "https://github.com/ethanabrooks/dollar-lambda.git", rev = "main"}
transformers = "^4.20.1"
requests-toolbelt = "^0.9.1"
rich = "^12.5.1"

[tool.poetry.dev-dependencies]
ipython = "^8.2.0"
ipdb = "^0.13.9"
black = "^22.3.0"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
